# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PicGlot API",
    "description": "\nRecognise, translate, edit and convert text in images and documents.\n\n**Authentication** — session cookie for the web app, `Authorization: Bearer lik_live_…`\nfor the public API.\n\n**Idempotency** — send an `Idempotency-Key` header on any POST that creates a job or\na payment; repeating the request returns the original result instead of doing the work\ntwice.\n\n**Errors** — every failure returns `{\"error\": {\"code\", \"message\", \"retryable\"}}` with a\nstable machine-readable `code`.\n",
    "contact": {
      "name": "PicGlot",
      "email": "support@picglot.ru"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "development"
    }
  ],
  "paths": {
...
    "/api/v1/billing/checkout": {
      "post": {
        "tags": ["billing"],
        "summary": "Checkout",
        "operationId": "checkout_api_v1_billing_checkout_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/subscription": {
      "get": {
        "tags": ["billing"],
        "summary": "Current Subscription",
        "operationId": "current_subscription_api_v1_billing_subscription_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SubscriptionOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Current Subscription Api V1 Billing Subscription Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/subscription/cancel": {
      "post": {
        "tags": ["billing"],
        "summary": "Cancel Subscription",
        "operationId": "cancel_subscription_api_v1_billing_subscription_cancel_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/payments": {
      "get": {
        "tags": ["billing"],
        "summary": "List Payments",
        "operationId": "list_payments_api_v1_billing_payments_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentOut"
                  },
                  "title": "Response List Payments Api V1 Billing Payments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutOut": {
        "properties": {
          "checkout_url": {
            "type": "string",
            "title": "Checkout Url"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "test_mode": {
            "type": "boolean",
            "title": "Test Mode",
            "default": false
          }
        },
        "type": "object",
        "required": ["checkout_url", "provider"],
        "title": "CheckoutOut"
      },
      "CheckoutRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": ["subscription", "credit_pack"],
            "title": "Kind",
            "default": "subscription"
          },
          "plan_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Code"
          },
          "pack_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Code"
          },
          "coupon": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 48
              },
              {
                "type": "null"
              }
            ],
            "title": "Coupon"
          },
          "success_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success Url"
          },
          "cancel_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Url"
          }
        },
        "type": "object",
        "title": "CheckoutRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "amount_minor": {
            "type": "integer",
            "title": "Amount Minor"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "credits_granted": {
            "type": "integer",
            "title": "Credits Granted"
          },
          "invoice_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "amount_minor",
          "currency",
          "status",
          "kind",
          "credits_granted",
          "invoice_url",
          "created_at"
        ],
        "title": "PaymentOut"
      },
...
      "SubscriptionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "plan_code": {
            "type": "string",
            "title": "Plan Code"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "current_period_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period End"
          },
          "cancel_at_period_end": {
            "type": "boolean",
            "title": "Cancel At Period End"
          }
        },
        "type": "object",
        "required": [
          "id",
          "plan_code",
          "status",
          "provider",
          "current_period_end",
          "cancel_at_period_end"
        ],
        "title": "SubscriptionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}